feat: Move options and storage parameter classes under Table class
